requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.5.1
eventlet==0.39.1
paho-mqtt==2.1.0
PyMySQL==1.1.1
python-dotenv==1.0.1
Werkzeug==3.1.3
pytest==8.3.5
